precision: default to float32, pin the test suite to float64
Flip the package working precision default (mrx/precision.py) from float64 to
float32, the production precision, so a fresh import / the tutorials run in
float32 without setting MRX_DTYPE. The lean suite verifies at double precision
(manufactured Poisson, harmonic-form ratios), so conftest pins MRX_DTYPE to
float64 before mrx is imported -- its behaviour is unchanged (51/51 pass).
gvec.py no longer hardcodes float64 for the Fourier mode-number arrays (it
follows the working dtype), which otherwise truncates and warns on every read
under the float32 default.

+++ mrx/precision.py
@@ -1,7 +1,7 @@
 """Working precision of the package.
 
 The precision is chosen once, from the environment variable ``MRX_DTYPE``
-(``float64``, the default, or ``float32``), before any array is created.
+(``float32``, the default, or ``float64``), before any array is created.
 Importing :mod:`mrx` applies it; scripts and tests do not touch
 ``jax_enable_x64`` themselves.
 
@@ -17,7 +17,7 @@ import jax
 import jax.numpy as jnp
 import numpy as np
 
-_NAME = os.environ.get("MRX_DTYPE", "float64")
+_NAME = os.environ.get("MRX_DTYPE", "float32")
 if _NAME not in ("float32", "float64"):
     raise ValueError(
         f"MRX_DTYPE={_NAME!r}; expected 'float32' or 'float64'")
